requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
lxml==5.2.0
pydantic==2.6.4
pydantic_core==2.16.3
pypika-tortoise==0.1.6
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
sniffio==1.3.1
starlette==0.36.3
tortoise-orm==0.20.0
typing_extensions==4.10.0
uvicorn==0.29.0
